+++ src/types.rs
@@ -1,2 +1,2 @@
 #[allow(non_camel_case_types)] //its a primative
-pub type temp = isize;
+pub type temp = f32;
